# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Profy Skill Academy API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for the Profy Skill Academy learning platform (TECHNICAL_DOC §5).

    * Consumer endpoints (mobile + web share them) accept a JWT with audience
      `consumer`. Admin endpoints live under `/api/v1/admin/*` and accept a JWT
      with audience `admin`. The two token types are never interchangeable.
    * Errors use the envelope `{"error": {"code": string, "message": string}}`
      with stable machine codes (`ai_unavailable`, `rate_limited`, ...).
    * Clients send an `X-App-Version` header; the API can force-update later.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Create a consumer account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created; token pair returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in and receive a token pair
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token (reuse detection revokes the family)
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke the refresh token
      operationId: logout
      security:
        - consumerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ------------------------------------------------------------ public taxonomy
...
  /api/v1/lessons/{id}/ai/chat:
    post:
      tags: [ai]
      summary: Ask the AI Teacher about this lesson (session auto-created)
      operationId: aiChat
      security:
        - consumerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatRequest"
      responses:
        "200":
          description: Grounded reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIChatResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AIUnavailable"
  /api/v1/lessons/{id}/ai/messages:
    get:
      tags: [ai]
      summary: Chat history for this lesson (capped at 10 turns in prompts)
      operationId: aiMessages
      security:
        - consumerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ------------------------------------------------------- progress & library
...
  /api/v1/entitlement:
    get:
      tags: [billing]
      summary: Current entitlement (cached 60s)
      operationId: entitlement
      security:
        - consumerAuth: []
      responses:
        "200":
          description: Premium status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/billing/checkout:
    post:
      tags: [billing]
      summary: Create a Stripe Checkout session (web only)
      operationId: createCheckout
      security:
        - consumerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/admin/auth/login:
    post:
      tags: [admin-auth]
      summary: Admin login (issues JWT with audience `admin`)
      operationId: adminLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin token pair + profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/admin/auth/refresh:
    post:
      tags: [admin-auth]
      summary: Rotate the admin refresh token
      operationId: adminRefresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New admin token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/admin/auth/logout:
    post:
      tags: [admin-auth]
      summary: Revoke the admin refresh token
      operationId: adminLogout
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Revoked
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ------------------------------------------------------------ admin taxonomy
...
components:
  securitySchemes:
    consumerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Consumer JWT with audience `consumer` (15 min).
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT with audience `admin` (15 min). Never interchangeable with consumer tokens.
  responses:
    Unauthorized:
      description: Missing/invalid/expired token, or token audience mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Invalid request body/parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: "Too many requests (auth: IP-based; AI chat: per-user daily)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    AIUnavailable:
      description: "AI Teacher degraded - lesson content stays fully usable"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code, e.g. ai_unavailable, rate_limited.
            message:
              type: string
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
        device:
          type: string
          enum: [mobile, web]
          default: web
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        device:
          type: string
          enum: [mobile, web]
          default: web
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime in seconds (900).

    # ------------------------------------------------------------- taxonomy
...
    AIChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
    AIChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        created_at:
          type: string
          format: date-time
    ChatHistory:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"

    # ------------------------------------------------------------- progress
...
    Entitlement:
      type: object
      properties:
        is_premium:
          type: boolean
        plan:
          type: [string, "null"]
          enum: [monthly, yearly]
        current_period_end:
          type: [string, "null"]
          format: date-time
    CheckoutRequest:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [monthly, yearly]
        success_url:
          type: string
        cancel_url:
          type: string
    CheckoutResponse:
      type: object
      properties:
        checkout_url:
          type: string
        session_id:
          type: string

    # ----------------------------------------------------------- admin auth
    AdminLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin]
    AdminLoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        admin_user:
          $ref: "#/components/schemas/AdminUser"

    # ------------------------------------------------------- admin content
...